Move MockFileSource to fixtures
It is going to be used by another test that needs to emulate
asynchronous requests.
